refactor: split cmd/mnemo by command domain
main.go mixed memory, init, MCP, hook JSON helpers, and
install-instructions. Setup, doctor, and projects already
lived in their own files.

Keep package main; main.go only dispatches and prints usage.
No behavior change.

Verified: go build ./..., go test ./...
