+++ cmd/mnemo/main.go
@@ -1,22 +1,12 @@
 package main
 
 import (
-	"bufio"
 	_ "embed"
-	"encoding/json"
 	"fmt"
-	"io"
 	"os"
-	"path/filepath"
-	"strconv"
-	"strings"
 	"text/template"
 
-	"github.com/jmeiracorbal/mnemo/internal/agentinit"
-	"github.com/jmeiracorbal/mnemo/internal/jsonmerge"
-	mcpserver "github.com/jmeiracorbal/mnemo/internal/mcp"
 	"github.com/jmeiracorbal/mnemo/internal/store"
-	"github.com/mark3labs/mcp-go/server"
 )
 
 var version = "dev"
@@ -104,705 +94,6 @@ func main() {
 		os.Exit(1)
 	}
 }
-
-// ─── mcp ─────────────────────────────────────────────────────────────────────
-
-func runMCP(s *store.Store) {
-	tools := ""
-	for _, arg := range os.Args[2:] {
-		if len(arg) > 8 && arg[:8] == "--tools=" {
-			tools = arg[8:]
-		}
-	}
-
-	allowlist := mcpserver.ResolveTools(tools)
-	srv, err := mcpserver.NewServerWithTools(s, version, allowlist)
-	if err != nil {
-		fmt.Fprintf(os.Stderr, "mnemo: mcp server: %v\n", err)
-		os.Exit(1)
-	}
-
-	if err := server.ServeStdio(srv); err != nil {
-		fmt.Fprintf(os.Stderr, "mnemo: mcp server error: %v\n", err)
-		os.Exit(1)
-	}
-}
-
-// ─── save ────────────────────────────────────────────────────────────────────
-
-func runSave(s *store.Store) {
-	args := os.Args[2:]
-	if len(args) < 2 {
-		fmt.Fprintln(os.Stderr, "usage: mnemo save <title> <content> [--type TYPE] [--project PROJECT] [--scope SCOPE] [--topic TOPIC_KEY]")
-		os.Exit(1)
-	}
-
-	title := args[0]
-	content := args[1]
-	typ := "manual"
-	project := ""
-	scope := ""
-	topicKey := ""
-
-	for i := 2; i < len(args)-1; i++ {
-		switch args[i] {
-		case "--type":
-			typ = args[i+1]
-			i++
-		case "--project":
-			project = args[i+1]
-			i++
-		case "--scope":
-			scope = args[i+1]
-			i++
-		case "--topic":
-			topicKey = args[i+1]
-			i++
-		}
-	}
-
-	sessionID := "manual-save"
-	if project != "" {
-		sessionID = "manual-save-" + project
-	}
-	if err := s.CreateSession(sessionID, project, ""); err != nil {
-		fmt.Fprintf(os.Stderr, "mnemo: warning: could not create session: %v\n", err)
-	}
-
-	id, err := s.AddObservation(store.AddObservationParams{
-		SessionID: sessionID,
-		Type:      typ,
-		Title:     title,
-		Content:   content,
-		Project:   project,
-		Scope:     scope,
-		TopicKey:  topicKey,
-	})
-	if err != nil {
-		fmt.Fprintf(os.Stderr, "mnemo: save failed: %v\n", err)
-		os.Exit(1)
-	}
-
-	fmt.Printf("Memory saved: #%d %q (%s)\n", id, title, typ)
-}
-
-// ─── search ──────────────────────────────────────────────────────────────────
-
-func runSearch(s *store.Store) {
-	args := os.Args[2:]
-	if len(args) == 0 {
-		fmt.Fprintln(os.Stderr, "usage: mnemo search <query> [--type TYPE] [--project PROJECT] [--scope SCOPE] [--limit N]")
-		os.Exit(1)
-	}
-
-	query := args[0]
-	opts := store.SearchOptions{Limit: 10}
-
-	for i := 1; i < len(args)-1; i++ {
-		switch args[i] {
-		case "--type":
-			opts.Type = args[i+1]
-			i++
-		case "--project":
-			opts.Project = args[i+1]
-			i++
-		case "--scope":
-			opts.Scope = args[i+1]
-			i++
-		case "--limit":
-			if n, err := strconv.Atoi(args[i+1]); err == nil {
-				opts.Limit = n
-			}
-			i++
-		}
-	}
-
-	results, err := s.Search(query, opts)
-	if err != nil {
-		fmt.Fprintf(os.Stderr, "mnemo: search failed: %v\n", err)
-		os.Exit(1)
-	}
-
-	if len(results) == 0 {
-		fmt.Printf("No memories found for: %q\n", query)
-		return
-	}
-
-	fmt.Printf("Found %d memories:\n\n", len(results))
-	for i, r := range results {
-		proj := ""
-		if r.Project != nil {
-			proj = " | project: " + *r.Project
-		}
-		content := r.Content
-		if len(content) > 300 {
-			content = content[:300] + "... [truncated — use mem_get_observation(id) for full content]"
-		}
-		fmt.Printf("[%d] #%d (%s) — %s\n    %s\n    %s%s | scope: %s\n\n",
-			i+1, r.ID, r.Type, r.Title, content, r.CreatedAt, proj, r.Scope)
-	}
-}
-
-// ─── context ─────────────────────────────────────────────────────────────────
-
-func runContext(s *store.Store) {
-	project := ""
-	if len(os.Args) > 2 {
-		project = os.Args[2]
-	}
-
-	ctx, err := s.FormatContext(project, "")
-	if err != nil {
-		fmt.Fprintf(os.Stderr, "mnemo: context failed: %v\n", err)
-		os.Exit(1)
-	}
-
-	if ctx == "" {
-		fmt.Println("No previous session memories found.")
-		return
-	}
-
-	fmt.Println(ctx)
-}
-
-// ─── session ─────────────────────────────────────────────────────────────────
-
-func runSession(s *store.Store) {
-	if len(os.Args) < 4 {
-		fmt.Fprintln(os.Stderr, "usage: mnemo session start <id> [--project PROJECT] [--dir DIR]")
-		fmt.Fprintln(os.Stderr, "       mnemo session end <id> [--summary SUMMARY]")
-		os.Exit(1)
-	}
-
-	subcmd := os.Args[2]
-	id := os.Args[3]
-	args := os.Args[4:]
-
-	switch subcmd {
-	case "start":
-		project := ""
-		dir := ""
-		for i := 0; i < len(args)-1; i++ {
-			switch args[i] {
-			case "--project":
-				project = args[i+1]
-				i++
-			case "--dir":
-				dir = args[i+1]
-				i++
-			}
-		}
-		if err := s.CreateSession(id, project, dir); err != nil {
-			fmt.Fprintf(os.Stderr, "mnemo: session start failed: %v\n", err)
-			os.Exit(1)
-		}
-		fmt.Printf("Session %q started\n", id)
-
-	case "end":
-		summary := ""
-		for i := 0; i < len(args)-1; i++ {
-			if args[i] == "--summary" {
-				summary = args[i+1]
-				i++
-			}
-		}
-		if err := s.EndSession(id, summary); err != nil {
-			fmt.Fprintf(os.Stderr, "mnemo: session end failed: %v\n", err)
-			os.Exit(1)
-		}
-		fmt.Printf("Session %q completed\n", id)
-
-	case "exists":
-		_, err := s.GetSession(id)
-		if err != nil {
-			fmt.Println("false")
-			os.Exit(1)
-		}
-		fmt.Println("true")
-
-	case "obs-count":
-		n, err := s.ObsCount(id)
-		if err != nil {
-			fmt.Fprintf(os.Stderr, "mnemo: obs-count failed: %v\n", err)
-			os.Exit(1)
-		}
-		fmt.Println(n)
-
-	case "project-obs-count":
-		n, err := s.ObsCountForSession(id)
-		if err != nil {
-			fmt.Fprintf(os.Stderr, "mnemo: project-obs-count failed: %v\n", err)
-			os.Exit(1)
-		}
-		fmt.Println(n)
-
-	default:
-		fmt.Fprintf(os.Stderr, "mnemo: unknown session subcommand %q\n", subcmd)
-		os.Exit(1)
-	}
-}
-
-// ─── stats ───────────────────────────────────────────────────────────────────
-
-func runStats(s *store.Store) {
-	stats, err := s.Stats()
-	if err != nil {
-		fmt.Fprintf(os.Stderr, "mnemo: stats failed: %v\n", err)
-		os.Exit(1)
-	}
-
-	fmt.Printf("Memory stats:\n")
-	fmt.Printf("  Sessions:     %d\n", stats.TotalSessions)
-	fmt.Printf("  Observations: %d\n", stats.TotalObservations)
-	fmt.Printf("  Prompts:      %d\n", stats.TotalPrompts)
-	if len(stats.Projects) > 0 {
-		fmt.Printf("  Projects:     %v\n", stats.Projects)
-	}
-}
-
-// ─── export ──────────────────────────────────────────────────────────────────
-
-func runExport(s *store.Store) {
-	data, err := s.Export()
-	if err != nil {
-		fmt.Fprintf(os.Stderr, "mnemo: export failed: %v\n", err)
-		os.Exit(1)
-	}
-
-	out, err := json.MarshalIndent(data, "", "  ")
-	if err != nil {
-		fmt.Fprintf(os.Stderr, "mnemo: json marshal failed: %v\n", err)
-		os.Exit(1)
-	}
-
-	if len(os.Args) > 2 {
-		if err := os.WriteFile(os.Args[2], out, 0644); err != nil {
-			fmt.Fprintf(os.Stderr, "mnemo: write failed: %v\n", err)
-			os.Exit(1)
-		}
-		fmt.Printf("Exported to %s\n", os.Args[2])
-	} else {
-		fmt.Println(string(out))
-	}
-}
-
-// ─── import ──────────────────────────────────────────────────────────────────
-
-func runImport(s *store.Store) {
-	if len(os.Args) < 3 {
-		fmt.Fprintln(os.Stderr, "usage: mnemo import <file.json>")
-		os.Exit(1)
-	}
-
-	data, err := os.ReadFile(os.Args[2])
-	if err != nil {
-		fmt.Fprintf(os.Stderr, "mnemo: read failed: %v\n", err)
-		os.Exit(1)
-	}
-
-	var payload store.ExportData
-	if err := json.Unmarshal(data, &payload); err != nil {
-		fmt.Fprintf(os.Stderr, "mnemo: invalid json: %v\n", err)
-		os.Exit(1)
-	}
-
-	result, err := s.Import(&payload)
-	if err != nil {
-		fmt.Fprintf(os.Stderr, "mnemo: import failed: %v\n", err)
-		os.Exit(1)
-	}
-
-	fmt.Printf("Import complete: %d sessions, %d observations\n", result.SessionsImported, result.ObservationsImported)
-}
-
-// ─── capture ─────────────────────────────────────────────────────────────────
-
-func runCapture(s *store.Store) {
-	args := os.Args[2:]
-	if len(args) == 0 {
-		fmt.Fprintln(os.Stderr, "usage: mnemo capture <content>|- [--session SESSION_ID] [--project PROJECT]")
-		os.Exit(1)
-	}
-
-	var content string
-	if args[0] == "-" {
-		data, err := io.ReadAll(os.Stdin)
-		if err != nil {
-			fmt.Fprintf(os.Stderr, "mnemo: failed to read stdin: %v\n", err)
-			os.Exit(1)
-		}
-		content = string(data)
-	} else {
-		content = args[0]
-	}
-	sessionID := ""
-	project := ""
-
-	for i := 1; i < len(args)-1; i++ {
-		switch args[i] {
-		case "--session":
-			sessionID = args[i+1]
-			i++
-		case "--project":
-			project = args[i+1]
-			i++
-		}
-	}
-
-	if sessionID == "" {
-		sessionID = "manual-save"
-		if project != "" {
-			sessionID = "manual-save-" + project
-		}
-	}
-	if err := s.CreateSession(sessionID, project, ""); err != nil {
-		fmt.Fprintf(os.Stderr, "mnemo: warning: could not create session: %v\n", err)
-	}
-
-	result, err := s.PassiveCapture(store.PassiveCaptureParams{
-		SessionID: sessionID,
-		Content:   content,
-		Project:   project,
-		Source:    "subagent-stop",
-	})
-	if err != nil {
-		fmt.Fprintf(os.Stderr, "mnemo: capture failed: %v\n", err)
-		os.Exit(1)
-	}
-
-	fmt.Printf("Captured: extracted=%d saved=%d duplicates=%d\n",
-		result.Extracted, result.Saved, result.Duplicates)
-}
-
-// ─── json ────────────────────────────────────────────────────────────────────
-
-// runJSON reads JSON from stdin and extracts a value by key path.
-// Intended for use in hook scripts as a replacement for inline python3 -c.
-//
-// Usage: mnemo json KEY [KEY ...]
-// Examples:
-//
-//	mnemo json session_id
-//	mnemo json workspace_roots 0
-//	mnemo json tool_info transcript_path
-func runJSON() {
-	if len(os.Args) < 3 {
-		fmt.Fprintln(os.Stderr, "usage: mnemo json KEY [KEY ...]")
-		os.Exit(1)
-	}
-
-	var data any
-	if err := json.NewDecoder(os.Stdin).Decode(&data); err != nil {
-		fmt.Println("")
-		return
-	}
-
-	value := data
-	for _, key := range os.Args[2:] {
-		switch v := value.(type) {
-		case map[string]any:
-			value = v[key]
-		case []any:
-			i, err := strconv.Atoi(key)
-			if err != nil || i < 0 || i >= len(v) {
-				value = nil
-			} else {
-				value = v[i]
-			}
-		default:
-			value = nil
-		}
-		if value == nil {
-			break
-		}
-	}
-
-	switch v := value.(type) {
-	case string:
-		fmt.Println(v)
-	case float64:
-		if v == float64(int64(v)) {
-			fmt.Println(int64(v))
-		} else {
-			fmt.Println(v)
-		}
-	case bool:
-		fmt.Println(v)
-	default:
-		fmt.Println("")
-	}
-}
-
-// ─── extract-transcript ───────────────────────────────────────────────────────
-
-// runExtractTranscript reads a JSONL transcript file and prints the text
-// content of all assistant messages. Intended for passive capture in hooks.
-//
-// Usage: mnemo extract-transcript /path/to/transcript.jsonl
-func runExtractTranscript() {
-	if len(os.Args) < 3 {
-		fmt.Fprintln(os.Stderr, "usage: mnemo extract-transcript <file.jsonl>")
-		os.Exit(1)
-	}
-
-	f, err := os.Open(os.Args[2])
-	if err != nil {
-		// Silent failure: transcript may not exist yet.
-		return
-	}
-	defer f.Close()
-
-	type block struct {
-		Type string `json:"type"`
-		Text string `json:"text"`
-	}
-	type message struct {
-		Role    string          `json:"role"`
-		Content json.RawMessage `json:"content"`
-	}
-
-	var lines []string
-	scanner := bufio.NewScanner(f)
-	scanner.Buffer(make([]byte, 1024*1024), 1024*1024)
-	for scanner.Scan() {
-		var msg message
-		if err := json.Unmarshal(scanner.Bytes(), &msg); err != nil || msg.Role != "assistant" {
-			continue
-		}
-
-		// Try plain string content first.
-		var text string
-		if err := json.Unmarshal(msg.Content, &text); err == nil {
-			if text != "" {
-				lines = append(lines, text)
-			}
-			continue
-		}
-
-		// Try array of content blocks.
-		var blocks []block
-		if err := json.Unmarshal(msg.Content, &blocks); err == nil {
-			for _, b := range blocks {
-				if b.Type == "text" && b.Text != "" {
-					lines = append(lines, b.Text)
-				}
-			}
-		}
-	}
-
-	fmt.Println(strings.Join(lines, "\n"))
-}
-
-// ─── install instructions ───────────────────────────────────────────────────
-
-func runInstallInstructions() {
-	agent, home, err := parseInstallInstructionsArgs(os.Args[2:], os.UserHomeDir)
-	if err != nil {
-		fmt.Fprintf(os.Stderr, "mnemo install-instructions: home: %v\n", err)
-		os.Exit(1)
-	}
-
-	agents, err := agentinit.ExpandAgents(agent)
-	if err != nil {
-		fmt.Fprintf(os.Stderr, "mnemo install-instructions: %v\n", err)
-		os.Exit(1)
-	}
-	for _, a := range agents {
-		path, err := agentinit.InstallGlobalInstructions(home, a)
-		if err != nil {
-			fmt.Fprintf(os.Stderr, "mnemo install-instructions: %s: %v\n", a, err)
-			os.Exit(1)
-		}
-		fmt.Printf("mnemo install-instructions: %s updated %s\n", a, path)
-	}
-}
-
-func parseInstallInstructionsArgs(args []string, userHomeDir func() (string, error)) (agent, home string, err error) {
-	agent = "all"
-	for _, arg := range args {
-		switch {
-		case strings.HasPrefix(arg, "--agent="):
-			agent = arg[len("--agent="):]
-		case strings.HasPrefix(arg, "--home="):
-			home = arg[len("--home="):]
-		}
-	}
-	if home != "" {
-		return agent, home, nil
-	}
-	home, err = userHomeDir()
-	return agent, home, err
-}
-
-// ─── init ────────────────────────────────────────────────────────────────────
-
-// runInit activates mnemo for one or more agents in the current project.
-func runInit(s *store.Store) {
-	agent := "claudecode"
-	dir := "."
-
-	for _, arg := range os.Args[2:] {
-		switch {
-		case strings.HasPrefix(arg, "--agent="):
-			agent = arg[len("--agent="):]
-		case strings.HasPrefix(arg, "--path="):
-			dir = arg[len("--path="):]
-		}
-	}
-
-	abs, err := absPath(dir)
-	if err != nil {
-		fmt.Fprintf(os.Stderr, "mnemo init: %v\n", err)
-		os.Exit(1)
-	}
-	root := agentinit.ProjectRoot(abs)
-
-	projectID, err := agentinit.EnsureProjectID(root)
-	if err != nil {
-		fmt.Fprintf(os.Stderr, "mnemo init: project ID: %v\n", err)
-		os.Exit(1)
-	}
-	name := filepath.Base(root)
-	if err := s.EnsureProject(projectID, name); err != nil {
-		fmt.Fprintf(os.Stderr, "mnemo init: register project: %v\n", err)
-		os.Exit(1)
-	}
-
-	agents, err := agentinit.ExpandAgents(agent)
-	if err != nil {
-		fmt.Fprintf(os.Stderr, "mnemo init: %v\n", err)
-		os.Exit(1)
-	}
-
-	for _, a := range agents {
-		if err := agentinit.AddAgent(root, a); err != nil {
-			fmt.Fprintf(os.Stderr, "mnemo init: %s: %v\n", a, err)
-			os.Exit(1)
-		}
-		fmt.Printf("mnemo init: %s activated in %s\n", a, root)
-	}
-	if err := agentinit.EnsureGitignore(root); err != nil {
-		fmt.Fprintf(os.Stderr, "mnemo init: gitignore: %v\n", err)
-		os.Exit(1)
-	}
-	fmt.Printf("mnemo init: project ID %s\n", projectID)
-	if home, err := os.UserHomeDir(); err == nil {
-		printSkillRecommendation(os.Stdout, home)
-	}
-}
-
-func printSkillRecommendation(w io.Writer, home string) {
-	if agentinit.GlobalSkillInstalled(home) {
-		return
-	}
-	_, _ = fmt.Fprintln(w, "mnemo init: for a better agent experience, install the global mnemo-memory skill:")
-	_, _ = fmt.Fprintln(w, "  npx skills add jmeiracorbal/mnemo --skill mnemo-memory --global")
-}
-
-// runMigrateProjects migrates a project from a legacy path-derived key to a
-// UUID-based ID. Run once per project after upgrading to project-ID-based tracking.
-//
-// Steps:
-//  1. Resolve the current project root.
-//  2. Compute the deterministic UUID v5 from the absolute path.
-//  3. Compute what the legacy derived key would have been for this path.
-//  4. Rekey existing observations/sessions from legacy key → UUID.
-//  5. Register the project in the projects table.
-//  6. Write the UUID to .mnemo.
-func runMigrateProjects(s *store.Store) {
-	dir := "."
-	for _, arg := range os.Args[2:] {
-		if strings.HasPrefix(arg, "--path=") {
-			dir = arg[len("--path="):]
-		}
-	}
-
-	abs, err := absPath(dir)
-	if err != nil {
-		fmt.Fprintf(os.Stderr, "mnemo migrate: %v\n", err)
-		os.Exit(1)
-	}
-	root := agentinit.ProjectRoot(abs)
-
-	legacyRoot := root
-	if resolved, err := filepath.EvalSymlinks(root); err == nil {
-		legacyRoot = resolved
-	}
-
-	projectID := agentinit.ProjectUUIDFromPath(root)
-	legacyKey := deriveLegacyKey(legacyRoot)
-	name := filepath.Base(root)
-
-	result, err := s.MigrateProject(legacyKey, projectID)
-	if err != nil {
-		fmt.Fprintf(os.Stderr, "mnemo migrate: rekey: %v\n", err)
-		os.Exit(1)
-	}
-
-	if err := s.EnsureProject(projectID, name); err != nil {
-		fmt.Fprintf(os.Stderr, "mnemo migrate: register: %v\n", err)
-		os.Exit(1)
-	}
-
-	if err := agentinit.EnsureMarkerWithID(root, projectID); err != nil {
-		fmt.Fprintf(os.Stderr, "mnemo migrate: write marker: %v\n", err)
-		os.Exit(1)
-	}
-
-	if err := agentinit.EnsureGitignore(root); err != nil {
-		fmt.Fprintf(os.Stderr, "mnemo migrate: gitignore: %v\n", err)
-		os.Exit(1)
-	}
-
-	if result.Migrated {
-		fmt.Printf("migrated: %s -> %s (%d obs, %d sessions)\n",
-			legacyKey, projectID, result.ObservationsUpdated, result.SessionsUpdated)
-	} else {
-		fmt.Printf("no legacy data found for %s, project registered as %s\n", legacyKey, projectID)
-	}
-}
-
-// deriveLegacyKey computes what the path-derived project key would have been
-// for the given absolute path under the old hook derivation logic:
-//
-//	realpath | sed "s|^$HOME/||; s|^/||" | tr '/' '-' | tr '[:upper:]' '[:lower:]'
-func deriveLegacyKey(absPath string) string {
-	home, _ := os.UserHomeDir()
-	rel := absPath
-	if home != "" && strings.HasPrefix(rel, home+"/") {
-		rel = rel[len(home)+1:]
-	} else if strings.HasPrefix(rel, "/") {
-		rel = rel[1:]
-	}
-	return strings.ToLower(strings.ReplaceAll(rel, "/", "-"))
-}
-
-func absPath(dir string) (string, error) {
-	return filepath.Abs(dir)
-}
-
-// ─── json-merge ──────────────────────────────────────────────────────────────
-
-// runJSONMerge reads a JSON patch from stdin and deep-merges it into FILE.
-// Usage: mnemo json-merge <file>
-func runJSONMerge() {
-	if len(os.Args) < 3 {
-		fmt.Fprintln(os.Stderr, "usage: mnemo json-merge <file>")
-		os.Exit(1)
-	}
-	changed, err := jsonmerge.MergeFile(os.Args[2])
-	if err != nil {
-		fmt.Fprintf(os.Stderr, "mnemo: json-merge: %v\n", err)
-		os.Exit(1)
-	}
-	if changed {
-		fmt.Println("updated")
-	} else {
-		fmt.Println("already up to date")
-	}
-}
-
-// ─── usage ───────────────────────────────────────────────────────────────────
 
 func printUsage() {
 	if err := usageTemplate.Execute(os.Stderr, usageTemplateData{Version: version}); err != nil {
